feat: move pages to appdir
